move ModelFactory to a seperate file
